requirements.txt
fastapi
pydantic
uvicorn
SQLAlchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
isort
flake8
black
llama-index
pypdf
